Syntax for pre/post selector
Syntax if #rg?name where `name` is either `pre` or `post` and
`rg` is a list of extended ranges of the form, each of them
of the form:

  [slice1,slice2,...] or [/ slice1,slice2,...]

where `sliceX` are either integers or pair (i:j) of integers.

This syntax is meant to be a copy of resp. the pre/post
of the active goal. Extended ranges allow to copy only
a sub-part of the pre/post by seeing the formula as a
big conjunction and selecting the sub-formulas matching the
extended ranges. The [/] variant of the extended range does
a recursive flattening of the formula, while its raw form []
only flattens the top-level /\.